Switch to a simple array implementation and implemented undo
This is unfortunate, because we lose some of the power of observables. I was trying it out because I can't get observables to allow for dropping the last item, which is what we need for undo. Using .scan in conjunction with an operators array was getting in the way more than anything else. The best alternative I could consider is using a stream of the current operators so the sum can recalculate automatically.